refactor(tests): consolidate duplicated blob-offload integration fixtures into shared helpers (#4987)
The blob-offload trace read-path tests copied fixture helpers verbatim across
files, so the ClickHouse row/event shapes were encoded in several places and a
schema change had to be applied N times (the reuse-first rule the repo CLAUDE.md
calls out).

Extract the helpers that are byte-identical across the two primary files
(trace-detail-blob-recall-endpoint + large-trace-blob-offload-readpath) into a
shared module blob-offload-test-helpers.ts: AGGREGATE_TYPE, UNIQUE_TAIL,
LARGE_VALUE, assertOverThreshold, insertEventLogRow, extractSpanAttrs. Extract
the openProtections constant (verified byte-identical across 5 files) into
open-protections.ts and import it in all five.

Helpers that differ across files are deliberately left in place:
makeSpanReceivedEvent has divergent signatures per file, and the
insertTraceSummary/insertSpan/makeSpanRow family overlaps cross-directory with
real ClickHouse-row-shape drift risk. Consolidating those is out of scope here
to avoid silently changing fixture behavior.

Behavior-preserving, test-only.

Closes #4987
